refactoring sample survey and basic infog
