+++ packages/main-ui/src/pages/deliberations/new/details/sample_survey/components/mod.rs
@@ -1,0 +1,3 @@
+pub mod introduction;
+pub mod member;
+pub mod reward;
