+++ packages/main-ui/src/pages/deliberations/new/details/sample_survey/mod.rs
@@ -1,5 +1,7 @@
+mod components;
 mod controller;
 mod i18n;
 mod page;
 
+pub(self) use super::DeliberationNewController;
 pub use page::*;
